Deploy script for staging (#18)
We have many contracts now, we need a more streamlined deploy and setup
process. I wrote a class called `NeokingdomDAO` that takes care of this.

It has two instances:

- `NeokingdomDAOHardhat` use it together with HardHat tasks, it outputs
logs in the console, and it stores things on disk.
- `NeokingdomDAOMemory` use it for testing, it keeps everything in
memory.

The class is used to run "sequences", generic list of commands to do
things like deploying the DAO or setting it up.

What is pretty nice is that when running a sequence using
`NeokingdomDAOHardhat`, the class stores the progress in a file. In case
of an exception (i.e. when there is a network issue and the transaction
fails), if the sequence is run again it will start from the last
checkpoint.

This is still quite experimental, but if you look at the deploy sequence
you'll see how clean and simple the code is:
https://github.com/NeokingdomDAO/contracts/pull/18/files#diff-52b83d9606936f7d0a369cfdb236d7bdaf3861538f4cb4c35d0843445cd71864R138-R207

Closes #12, closes #8

---------

Co-authored-by: Nicola Miotto <miotto@posteo.de>